refactor(docs): divider dx
